requirements.txt
# parsing
beautifulsoup4==4.8.0

# Celery
celery==4.4.0
colormath==3.0.0
colorthief==0.2.1
# Django
Django==2.2.12 #4.1.7
django-allauth==0.41.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-extensions==2.2.1
djangorestframework==3.10.3
drf-yasg==1.17.0
environs==7.1.0
fake-useragent==0.1.11
google-cloud-vision==0.41.0

# Http
gunicorn==19.9.0
html2text==2019.9.26

# TinEye
# infrastructure/libraries/pytineye/pytineye.zip
lxml==4.4.2 #4.9.2
# opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30 #4.7.0.72
pandas==0.25.3  #1.5.3

# Other
pathlib==1.0.1
piexif==1.1.3

# Image
Pillow==6.2.1
# psycopg2-binary==2.8.4
pygeoip==0.3.2

#DataScience
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pymorphy2-dicts-ru==2.4.404381.4453942

recommonmark==0.6.0
requests==2.22.0
scikit-learn==0.22.1


# Logging
sentry-sdk==0.13.5

# Auto-doc
Sphinx==2.2.0
sphinx-rtd-theme==0.4.3
tldextract==2.2.2

# Analytics
ua-parser==0.8.0
user-agents==2.0

# API
vk-api==11.7.0
